Create content editor component
