+++ src/App.tsx
@@ -18,6 +18,7 @@ import { IoCopySharp } from "react-icons/io5";
 import TextEnhancementAction from "./components/TextEnhancementAction";
 import TextEnhancementActions from "./components/TextEnhancementActions";
 import Sidebar from "./components/Sidebar";
+import ContentEditor from "./components/ContentEditor";
 
 const Layout = styled.div`
   display: grid;
@@ -60,10 +61,6 @@ const MainWrapper = styled.main`
   overflow-y: auto;
 `;
 
-const getListStyle = (isDraggingOver: boolean) => ({
-  background: isDraggingOver ? "lightblue" : "white",
-});
-
 function App() {
   /* Todo:
     1. Integrate react-beautiful-dnd
@@ -73,70 +70,6 @@ function App() {
     5. Change the layout design 
   */
 
-  // This takes care of managing the focus on creation of new editable content row
-  const [newlyCreatedContentRowId, setNewlyCreatedContentRowId] = useState("");
-  const [contentRows, setContentRows] = useState<IContentRow[]>([
-    {
-      id: "item-1",
-      htmlContent: `<p>item-1 content</p>`,
-    },
-    {
-      id: "item-2",
-      htmlContent: `<p>item-2 content</p>`,
-    },
-    {
-      id: "item-3",
-      htmlContent: `<p>item-3 content</p>`,
-    },
-    {
-      id: "item-4",
-      htmlContent: `<p>item-4 content</p>`,
-    },
-    {
-      id: "item-5",
-      htmlContent: `<p>item-5 content</p>`,
-    },
-    {
-      id: "item-6",
-      htmlContent: `<p>item-6 content</p>`,
-    },
-  ]);
-
-  function onDragEnd(result: DropResult) {
-    // dropped outside the list
-    if (!result.destination) {
-      return;
-    }
-
-    const contentRowList = reorder(
-      contentRows,
-      result.source.index,
-      result.destination.index
-    );
-
-    setContentRows(contentRowList);
-  }
-  // Todo: needs to be moved to utilities
-  function handleAddContentEditableRow(contentRow: IContentRow) {
-    const createdNewContentRow: IContentRow = {
-      id: `item-${contentRows.length + 1}`,
-      htmlContent: `<p> </p>`,
-    };
-    setNewlyCreatedContentRowId(createdNewContentRow.id);
-
-    let contentRowList = Array.from(contentRows);
-    const currentClickedContentRowIndex = contentRows.findIndex(
-      (contentRowVal) => contentRowVal.id === contentRow.id
-    );
-    contentRowList.splice(
-      currentClickedContentRowIndex + 1,
-      0,
-      createdNewContentRow
-    );
-
-    setContentRows([...contentRowList]);
-  }
-
   return (
     <Layout>
       <SidebarWrapper border="lightgrey">
@@ -144,40 +77,7 @@ function App() {
       </SidebarWrapper>
       <MainWrapper background="white">
         <TextEnhancementActions />
-        <DragDropContext onDragEnd={onDragEnd}>
-          <Droppable droppableId="droppable">
-            {(droppableProvided, droppableSnapshot) => (
-              <div
-                ref={droppableProvided.innerRef}
-                style={getListStyle(droppableSnapshot.isDraggingOver)}
-              >
-                {contentRows.map((contentRow, index) => (
-                  <Draggable
-                    key={contentRow.id}
-                    draggableId={contentRow.id}
-                    index={index}
-                  >
-                    {(draggableProvided, draggableSnapshot) => {
-                      return (
-                        <DraggableRow
-                          draggableProvided={draggableProvided}
-                          draggableSnapshot={draggableSnapshot}
-                          contentRow={contentRow}
-                          setContentRows={setContentRows}
-                          handleAddContentEditableRow={
-                            handleAddContentEditableRow
-                          }
-                          newlyCreatedContentRowId={newlyCreatedContentRowId}
-                        />
-                      );
-                    }}
-                  </Draggable>
-                ))}
-                {droppableProvided.placeholder}
-              </div>
-            )}
-          </Droppable>
-        </DragDropContext>
+        <ContentEditor />
       </MainWrapper>
     </Layout>
   );
